Update queue module to use 'notification' queue; add REPL support and integrate devtools
